src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
CacheControl==0.14.4
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudevents==1.9.0
cryptography==46.0.3
deprecation==2.1.0
filelock==3.20.0
firebase-functions==0.1.2
firebase_admin==7.1.0
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.8.0
functions-framework==3.9.2
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-cloud-tasks==2.20.0
google-crc32c==1.7.1
google-events==0.14.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
grpclib==0.4.8
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modal==1.2.2
modelscope==1.32.0
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
sigtools==4.0.1
sniffio==1.3.1
starlette==0.50.0
synchronicity==0.10.2
toml==0.10.2
tqdm==4.67.1
typer==0.20.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvicorn-worker==0.4.0
watchdog==6.0.0
watchfiles==1.1.1
Werkzeug==3.1.4
yarl==1.22.0
